move recode_region() over the dart_pipeline.util
