+++ src/dart_pipeline/metrics/era5/util.py
@@ -25,7 +25,7 @@ from geoglue.cds import (
     DatasetPool,
 )
 from geoglue.util import get_first_monday
-from geoglue.region import ZonedBaseRegion, AdministrativeLevel
+from geoglue.region import ZonedBaseRegion
 
 from ...paths import get_path
 from .list_metrics import VARIABLES
@@ -474,11 +474,3 @@ def standardized_precipitation(
         case "spei":
             return norm_spi.rename(var)
 
-
-def recode_region(
-    ds: xr.Dataset | xr.DataArray, region: AdministrativeLevel
-) -> xr.Dataset | xr.DataArray:
-    geom = region.read()
-    gid_lookup = geom.reset_index(drop=True)[region.pk].astype(str).to_numpy()
-
-    return ds.assign_coords(region=("region", gid_lookup))
